Avec calibration

+++ powerManager/powerManager.h
@@ -25,7 +25,7 @@ extern volatile uint16_t voltageReading;
 extern volatile uint16_t currentReading;
 extern volatile power_t peakPower;
 
-void powerManager_loop(uint8_t v, uint8_t i);
+void powerManager_loop(uint16_t v, uint16_t i);
 void powerManager_setMotorPower(int16_t leftPower, int16_t rightPower);
 
 #endif // _POWER_MANAGER_H_
